board
기능 구현완료
-글 작성 (POST)
-글 수정 (PUT)
-ID 값으로 글 호출 (GET)
-전체 글 호출 (GET)

+++ src/main/java/com/github/board/Repository/PostRepository.java
@@ -8,5 +8,5 @@ public interface PostRepository extends JpaRepository<PostEntity, Long> {
 //     CRUD (Create, Read, Update, Delete) 작업을 수행 가능
 
     // 특정 ID로 게시물 조회 - Optional : ID에 해당하는 게시물이 존재하지 않을 경우 예외를 방지할 수 있음
-    Optional<PostEntity> findById(Long id);
+    Optional<PostEntity> findById(Integer id);
 }
